# clist: Speedup functions