We used unnecessary function calls, defers, and extra mutexes.
These are not our friends for writing fast code in our libs.

`defer()` was removed because these functions only had one or two code paths, so the minute increase in code complexity seems justified to me. I do think defers should be used outside of libs or in libs whenever there are several code paths. 

Before:
```
BenchmarkDetaching-8   	 5000000	       263 ns/op	       0 B/op	       0 allocs/op
BenchmarkRemoved-8     	20000000	        98.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkPushBack-8    	 2000000	       637 ns/op	     328 B/op	       5 allocs/op
```
After:
```
BenchmarkDetaching-8   	20000000	       111 ns/op	       0 B/op	       0 allocs/op
BenchmarkRemoved-8     	20000000	        61.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkPushBack-8    	 3000000	       438 ns/op	     328 B/op	       5 allocs/op
```
Originally this commit was just for detaching, then I realized defers were used in the same fashion throughout the file. I think the benchmarks for detaching are indicative that this is a good idea, but if desired I can benchmark more of the functions. `BenchmarkRemoved` and `BenchmarkPushback` should be indicative of the gains experienced by most of the functions with this change. 

<!-- Thanks for filing a PR! Before hitting the button, please check the following items.-->

* [X] Updated all relevant documentation in docs - nothing changes as a user of the API
* [ ] Updated all code comments where relevant - anything that warrants a comment?
* [X] Wrote tests
* [ ] Updated CHANGELOG_PENDING.md - do we update changelog for this sort of stuff?